app/requirements.txt
alembic==1.0.10
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
facebook-sdk==3.1.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.4.0
Flask-OAuth==0.12
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==2.19.2
mysqlclient==1.4.2.post1
oauth2==1.9.0.post1
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
rauth==0.7.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.2
Werkzeug==0.15.4
